requirements.txt
Flask>=2.3.0,<3.0
spotipy>=2.23.0,<3.0
APScheduler>=3.9.0,<4.0
python-dotenv>=1.0.0,<2.0
pytz>=2023.3
pyOpenSSL>=25.0.0,<26.0
# Optional: For production deployment (choose one or similar)
# gunicorn>=20.1.0
# waitress>=2.1.0